payBilln't promo

+++ src/main/java/quanly/douong/entity/Promotion.java
@@ -14,7 +14,7 @@ import java.util.Date;
 public class Promotion {
     private String promotionId;
     private String promoName;
-    private Float discount;
+    private Double discount;
     private Date startDate;
     private Date endDate;
     private boolean status;
